Move helper funcs to power_tools
